+++ GrowingAutotracker/GrowingAutotracker.h
@@ -19,7 +19,6 @@
 
 #import "GrowingDynamicProxy.h"
 #import "GrowingAutotrackConfiguration.h"
-#import "GrowingRealAutotracker.h"
 #import <UIKit/UIKit.h>
 
 @interface GrowingAutotracker : GrowingDynamicProxy
