+++ GrowingAutotrackerCore/GrowingAutotrackConfiguration.h
@@ -22,6 +22,13 @@
 
 NS_ASSUME_NONNULL_BEGIN
 
+typedef NS_ENUM(NSUInteger, GrowingIgnorePolicy) {
+    GrowingIgnoreNone = 0,
+    GrowingIgnoreSelf = 1,     // 忽略自身
+    GrowingIgnoreChildren = 2, // 忽略所有子页面和孙子页面
+    GrowingIgnoreAll = 3,      // 忽略自身 + 忽略所有子页面和孙子页面
+};
+
 @interface GrowingAutotrackConfiguration : GrowingTrackConfiguration
 
 @property(nonatomic, assign) float impressionScale;
