+++ GrowingAutotrackerCore/GrowingNode/Category/UIViewController+GrowingNode.h
@@ -19,7 +19,6 @@
 
 
 #import <UIKit/UIKit.h>
-#import "GrowingRealAutotracker.h"
 #import "GrowingNodeProtocol.h"
 
 @interface UIViewController (GrowingNode) <GrowingNode>
@@ -27,7 +26,6 @@
 @end
 
 @interface UIViewController (GrowingPrivateAttributes)
-
 
 - (void)mergeGrowingAttributesPvar:(NSDictionary<NSString *,NSObject *> *)growingAttributesPvar;
 - (void)removeGrowingAttributesPvar:(NSString *)key;
