+++ GrowingAutotrackerCore/Page/UIViewController+GrowingPageHelper.m
@@ -22,8 +22,9 @@
 #import "UIViewController+GrowingPageHelper.h"
 #import "UIViewController+GrowingNode.h"
 #import "GrowingPage.h"
-#import "GrowingRealAutotracker.h"
+#import "GrowingAutotrackConfiguration.h"
 #import "GrowingPrivateCategory.h"
+
 static void *const GROWING_PAGE_OBJECT = "GROWING_PAGE_OBJECT";
 
 @implementation UIViewController (GrowingPageHelper)
